requirements.txt
psutil
pyembedded
paho-mqtt